tests - alter table, export params, listing refactoring
