+++ tests/Backend/Common/ExportParamsTest.php
@@ -24,7 +24,7 @@ class ExportParamsTest extends StorageApiTestCase
 
 	public function testInvalidExportFormat()
 	{
-		$importFile =  __DIR__ . '/../_data/languages.csv';
+		$importFile =  __DIR__ . '/../../_data/languages.csv';
 		$tableId = $this->_client->createTable($this->getTestBucketId(), 'languages', new CsvFile($importFile));
 
 		try {
@@ -39,17 +39,17 @@ class ExportParamsTest extends StorageApiTestCase
 
 	public function testTableFileExport()
 	{
-		$importFile =  __DIR__ . '/../_data/languages.csv';
+		$importFile =  __DIR__ . '/../../_data/languages.csv';
 		$tableId = $this->_client->createTable($this->getTestBucketId(), 'languages', new CsvFile($importFile));
 
-		$outputFile = __DIR__ . '/../_tmp/languagesExport.csv';
+		$outputFile = __DIR__ . '/../../_tmp/languagesExport.csv';
 		if (file_exists($outputFile)) {
 			unlink($outputFile);
 		}
 
 		// Full download
 		$this->_client->exportTable($tableId, $outputFile);
-		$this->assertFileEquals($importFile, $outputFile);
+		$this->assertLinesEqualsSorted(file_get_contents($importFile), file_get_contents($outputFile));
 		unlink($outputFile);
 
 		// Download with limit
@@ -63,7 +63,7 @@ class ExportParamsTest extends StorageApiTestCase
 
 	public function testTableExportParams()
 	{
-		$importFile =  __DIR__ . '/../_data/languages.csv';
+		$importFile =  __DIR__ . '/../../_data/languages.csv';
 		$tableId = $this->_client->createTable($this->getTestBucketId(self::STAGE_IN), 'languages', new CsvFile($importFile));
 
 		$originalFileLinesCount = exec("wc -l <" . escapeshellarg($importFile));
@@ -106,7 +106,7 @@ class ExportParamsTest extends StorageApiTestCase
 	 */
 	public function testTableExportFilters($exportOptions, $expectedResult)
 	{
-		$importFile =  __DIR__ . '/../_data/users.csv';
+		$importFile =  __DIR__ . '/../../_data/users.csv';
 		$tableId = $this->_client->createTable($this->getTestBucketId(), 'users', new CsvFile($importFile));
 		$this->_client->markTableColumnAsIndexed($tableId, 'city');
 
@@ -119,7 +119,7 @@ class ExportParamsTest extends StorageApiTestCase
 
 	public function testTableExportFilterShouldFailOnNonIndexedColumn()
 	{
-		$importFile =  __DIR__ . '/../_data/users.csv';
+		$importFile =  __DIR__ . '/../../_data/users.csv';
 		$tableId = $this->_client->createTable($this->getTestBucketId(), 'users', new CsvFile($importFile));
 
 		try {
@@ -134,7 +134,7 @@ class ExportParamsTest extends StorageApiTestCase
 
 	public function testTableExportShouldFailOnNonExistingColumn()
 	{
-		$importFile =  __DIR__ . '/../_data/users.csv';
+		$importFile =  __DIR__ . '/../../_data/users.csv';
 		$tableId = $this->_client->createTable($this->getTestBucketId(self::STAGE_IN), 'users', new CsvFile($importFile));
 
 		try {
@@ -150,7 +150,7 @@ class ExportParamsTest extends StorageApiTestCase
 
 	public function testTableExportColumnsParam()
 	{
-		$importFile =  __DIR__ . '/../_data/languages.csv';
+		$importFile =  __DIR__ . '/../../_data/languages.csv';
 		$tableId = $this->_client->createTable($this->getTestBucketId(), 'languages', new CsvFile($importFile));
 
 		$data = $this->_client->exportTable($tableId, null, array(
@@ -164,111 +164,9 @@ class ExportParamsTest extends StorageApiTestCase
 		$this->assertEquals("id", $firstRow[0]);
 	}
 
-	public function testTableExportAsyncColumnsParam()
-	{
-		$importFile =  __DIR__ . '/../_data/languages.csv';
-		$tableId = $this->_client->createTable($this->getTestBucketId(), 'languages', new CsvFile($importFile));
-
-		$results = $this->_client->exportTableAsync($tableId, array(
-			'columns' => array('id'),
-		));
-		$file = $this->_client->getFile($results['file']['id']);
-		$parsed = Client::parseCsv(file_get_contents($file['url']), false);
-		$firstRow = reset($parsed);
-
-		$this->assertCount(1, $firstRow);
-		$this->assertArrayHasKey(0, $firstRow);
-		$this->assertEquals("id", $firstRow[0]);
-	}
-
-	/**
-	 * @param $exportOptions
-	 * @param $expectedResult
-	 * @dataProvider tableExportFiltersData
-	 */
-	public function testTableExportAsyncMysql($exportOptions, $expectedResult)
-	{
-		$importFile =  __DIR__ . '/../_data/users.csv';
-		$tableId = $this->_client->createTable($this->getTestBucketId(), 'users', new CsvFile($importFile));
-		$this->_client->markTableColumnAsIndexed($tableId, 'city');
-
-		$results = $this->_client->exportTableAsync($tableId, $exportOptions);
-
-		$exportedFile = $this->_client->getFile($results['file']['id']);
-		$parsedData = Client::parseCsv(file_get_contents($exportedFile['url']), false);
-		array_shift($parsedData); // remove header
-
-		$this->assertArrayEqualsSorted($expectedResult, $parsedData, 0);
-	}
-
-	/**
-	 * @param $exportOptions
-	 * @param $expectedResult
-	 * @dataProvider tableExportFiltersData
-	 */
-	public function testTableExportAsyncRedshift($exportOptions, $expectedResult)
-	{
-		$importFile =  __DIR__ . '/../_data/users.csv';
-		$csvFile = new CsvFile($importFile);
-		$tableId = $this->_client->createTableAsync($this->getTestBucketId(self::STAGE_IN, self::BACKEND_REDSHIFT), 'users', $csvFile, array(
-			'columns' => $csvFile->getHeader(),
-		));
-
-		$results = $this->_client->exportTableAsync($tableId, array_merge($exportOptions, array(
-			'format' => 'raw',
-		)));
-
-		$exportedFile = $this->_client->getFile($results['file']['id'], (new \Keboola\StorageApi\Options\GetFileOptions())->setFederationToken(true));
-
-		$this->assertTrue($exportedFile['isSliced']);
-		$this->assertGreaterThan(0, $exportedFile['sizeBytes']);
-
-		$manifest = json_decode(file_get_contents($exportedFile['url']), true);
-
-		$s3Client = new \Aws\S3\S3Client([
-			'credentials' => [
-				'key' => $exportedFile['credentials']['AccessKeyId'],
-				'secret' => $exportedFile['credentials']['SecretAccessKey'],
-				'token' => $exportedFile['credentials']['SessionToken'],
-			],
-			'version' => 'latest',
-			'region' => $exportedFile['region']
-		]);
-		$s3Client->registerStreamWrapper();
-
-		$csv = "";
-		foreach ($manifest['entries'] as $filePart) {
-			$csv .= file_get_contents($filePart['url']);
-		}
-
-		$parsedData = Client::parseCsv($csv, false, "\t", "");
-		$this->assertArrayEqualsSorted($expectedResult, $parsedData, 0);
-
-		// Check S3 ACL and listing bucket
-		$s3Client = new \Aws\S3\S3Client([
-			'credentials' => [
-				'key' => $exportedFile['credentials']['AccessKeyId'],
-				'secret' => $exportedFile['credentials']['SecretAccessKey'],
-				'token' => $exportedFile['credentials']['SessionToken'],
-			],
-			'version' => 'latest',
-			'region' => $exportedFile['region']
-		]);
-		$bucket = $exportedFile["s3Path"]["bucket"];
-		$prefix = $exportedFile["s3Path"]["key"];
-		$objects = $s3Client->listObjects(array(
-			"Bucket" => $bucket,
-			"Prefix" => $prefix
-		));
-		$this->assertEquals(3, count($objects["Contents"]));
-		foreach($objects["Contents"] as $object) {
-			$this->assertStringStartsWith($prefix, $object["Key"]);
-		}
-	}
-
 	public function testTableExportAsyncCache()
 	{
-		$importFile =  __DIR__ . '/../_data/users.csv';
+		$importFile =  __DIR__ . '/../../_data/users.csv';
 		$tableId = $this->_client->createTable($this->getTestBucketId(), 'users', new CsvFile($importFile));
 		$this->_client->markTableColumnAsIndexed($tableId, 'city');
 
@@ -319,7 +217,7 @@ class ExportParamsTest extends StorageApiTestCase
 	 */
 	public function testTableExportAsyncPermissions()
 	{
-		$importFile =  __DIR__ . '/../_data/users.csv';
+		$importFile =  __DIR__ . '/../../_data/users.csv';
 		$tableId = $this->_client->createTable($this->getTestBucketId(), 'users', new CsvFile($importFile));
 		$this->_client->markTableColumnAsIndexed($tableId, 'city');
 
@@ -331,7 +229,7 @@ class ExportParamsTest extends StorageApiTestCase
 			$this->getTestBucketId() => 'read',
 		));
 		$newToken = $this->_client->getToken($newTokenId);
-		$client = new Keboola\StorageApi\Client(array(
+		$client = new \Keboola\StorageApi\Client(array(
 			'token' => $newToken['token'],
 			'url' => STORAGE_API_URL,
 		));
@@ -343,26 +241,6 @@ class ExportParamsTest extends StorageApiTestCase
 
 		$file = $client->getFile($results['file']['id']);
 		Client::parseCsv(file_get_contents($file['url']), false);
-	}
-
-	public function testTableExportAsyncGzip()
-	{
-		$importFile =  __DIR__ . '/../_data/users.csv';
-		$tableId = $this->_client->createTable($this->getTestBucketId(), 'users', new CsvFile($importFile));
-		$this->_client->markTableColumnAsIndexed($tableId, 'city');
-
-		$results = $this->_client->exportTableAsync($tableId, array(
-			'gzip' => true,
-		));
-
-		$exportedFile = $this->_client->getFile($results['file']['id']);
-		$parsedData = Client::parseCsv(gzdecode(file_get_contents($exportedFile['url'])), false);
-		array_shift($parsedData); // remove header
-
-		$expected = Client::parseCsv(file_get_contents($importFile), false);
-		array_shift($expected);
-
-		$this->assertArrayEqualsSorted($expected, $parsedData, 0);
 	}
 
 }
